+++ fp-library/src/classes/semimonad.rs
@@ -9,7 +9,7 @@ use crate::{Apply, kinds::*};
 /// If `x` has type `m a` and `f` has type `a -> m b`, then `bind(x, f)` has type `m b`,
 /// representing the result of executing `x` to get a value of type `a` and then
 /// passing it to `f` to get a computation of type `m b`.
-pub trait Semimonad: Kind_c3c3610c70409ee6 {
+pub trait Semimonad: Kind_cdc7cd43dac7585f {
 	/// Sequences two computations, allowing the second to depend on the value computed by the first.
 	///
 	/// This method chains two computations, where the second computation depends on the result of the first.